+++ src/app/detection/detection.component.ts
@@ -3,6 +3,7 @@ import { MatSnackBar } from '@angular/material';
 import { DetectionOptions, EigenFaceRecognizerOptions, Status } from 'face-command-common';
 import { FaceCommandClientService } from '../face-command-client.service';
 import { AppErrorHandler } from '../app-error-handler';
+import { RPCModels } from 'face-command-client';
 
 /**
  * Component that allows the user to control detection.
@@ -49,21 +50,25 @@ export class DetectionComponent implements OnInit {
   /**
    * Status changes that have occured since detection started.
    */
-  public statusChanges: Status[] = [];
+  public lastStatus: Status;
   
+  public formatStatusBrightness(status: Status) {
+    return Math.round(status.brightness*100);
+  }
+
   /**
    * Stops the detection session.
    */
   async stopDetection() {
     await this.client.detectionService.StopDetection();
-    this.statusChanges = [];
+    this.lastStatus = null;
   }
 
   /**
    * Attempts to start the detection session.
    */
   async startDetection() {
-    this.statusChanges = [];
+    this.lastStatus = null;
     await this.client.detectionService.StartDetection(this.detectionOptions);
   }
 
@@ -73,11 +78,13 @@ export class DetectionComponent implements OnInit {
    */
   async ngOnInit() {
     this.isDetectionRunning = await this.client.detectionService.IsDetectionRunning();
+    this.lastStatus = await this.client.detectionService.GetLastStatus();
+
     const recOptions = new EigenFaceRecognizerOptions((await this.client.configService.GetConfigValue("eigenFaceRecognizerOptions:components")), (await this.client.configService.GetConfigValue("eigenFaceRecognizerOptions:threshold")));
     this.detectionOptions = new DetectionOptions((await this.client.configService.GetConfigValue("imageCaptureFrequency")), recOptions);
-    
+
     this.client.detectionService.on("StatusChange", (status: Status) => {
-      this.statusChanges.unshift(status);
+      this.lastStatus = status;
     });
 
     this.client.detectionService.on("DetectionRunning", (running: boolean) => {